+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Support for Characters on the 4th NA DC dynamis
 -   Added two properties to Groups: Order, and Publicity
 -   Added command `/group edit` to edit the publicity and order of groups
+-   Added command `/gimme-roles`
 
 ### Changed
 
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `convertRoleListToSelectOptions` now has a description limit of 100 characters
 -   Command `/for-each-member-in-server update-roles` operates off of groups now
 -   Command `/group create name: string expressConfiguration: boolean` can optionally used to specify group's publicity and order when a group is created
+-   FFlogs roles in unordered groups will no longer remove roles that a person is not qualified for, only in ordered groups will a unqualifying fflogs role be removed
 
 ### Fixed